Rename columnResizer.tsx
