+++ jsapp/js/projects/projectsTable/columnResizer.tsx
